JetsonLocal/requirements.txt
fastapi==0.123.5
uvicorn[standard]==0.38.0
requests==2.32.5
pydantic==2.12.5
python-dotenv==1.1.1
psutil==7.0.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
websockets==15.0.1
SpeechRecognition==3.10.1
pyserial==3.5